Admin can delete users[related data]

+++ resources/views/dashboard.blade.php
@@ -15,10 +15,11 @@
                     @endif
 
                     You are logged in as <strong>ADMIN</strong>!
+                    {{-- <a href="{{route('admin/add_user')}}" class="btn btn-primary" style="float:right">Add User</a>  --}}
                     <div class="panel panel-default">
                     <!-- Default panel contents -->
                     <div class="panel-heading">User Management</div>
-
+                           
                     <!-- Table -->
                     <table class="table">
                     <thead>
@@ -33,7 +34,7 @@
                             <th>{{$user->name}}</th>
                             <th>{{$user->email}}</th>
                             <th><a href="/admin/{{$user->id}}/edit">Edit</a></th>
-                            <th><a href="/admin/user_destroy/{{$user->id}}">Delete</a></th>
+                            <th><a href="/admin/user_destroy/{{$user->id}}" onclick="if(confirm('Warning:- This user and all related projects and other data will be deleted.'));else{ return false}">Delete</a></th>
                             {{-- <th>
                                 <form action="{{ route('admin.destroy',$user->id)}}">
                                     {{method_field('DELETE')}}
